Add files via upload
This is a much better version. V2 features the use of 2d arrays, a concept I didn't know, but had known of. 2 mins of googling later, I was able to use it very well (I hope) in 2 cases, eliminating the use of the variables line1, line2 & line3, while also being able to easily add and delete custom win conditions with a simple 2d array of 6 numbers per array. 
Reduces if conditions heavily. Also allows me to check if the game is won with a single line instead of 8 lines (for 8 possible scenarios!)
Even with added functionality, the code only comes to 77 lines. Yes I know everything can be written in 1 line and lines shouldn't be how you measure code, blah, blah, blah, but it is a useful metric in this case. Showing how effective and efficient V2 is compared to V1. And how much more readable!

(Added basic comments in various areas for understandability. As well as a section at the bottom which acts as a to do list and stuff I learnt. Also, the winCondition[] array was inspired by the recipe json file in Minecraft where the pattern is shown in []. Instead of a pattern, I have given the coordinates.)

Disclaimer : I call lists arrays as I started off programming with HTML, CSS, JS. I know they are lists and are technically different, but it means the same to me.